Web page indexing system

+++ searchengine-master/src/main/java/searchengine/repository/LemmaRepository.java
@@ -1,9 +1,10 @@
 package searchengine.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import searchengine.entity.Lemma;
 
 @Repository
-public interface LemmaRepository extends JpaRepository<Lemma, Long> {
+public interface LemmaRepository extends JpaRepository<Lemma, Integer> {
     Lemma findByLemmaText(String lemmaText);
 }
