+++ searchengine-master/src/main/java/searchengine/repository/PageRepository.java
@@ -1,11 +1,11 @@
 package searchengine.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import org.springframework.stereotype.Repository;
+
 import searchengine.entity.Page;
 
 @Repository
-public interface PageRepository extends JpaRepository<Page, Long> {
+public interface PageRepository extends JpaRepository<Page, Integer> {
 
 }
